==> analysis_easifish/bydatasets/273LU_L23/00.viz-250606.ipynb <==
# Inputs
---

import numpy as np, zarr, nrrd
# # alignment functions
# from bigstream.align import alignment_pipeline
# from bigstream.transform import apply_transform
import tifffile

import seaborn as sns
import matplotlib.pyplot as plt

p      = '/u/home/f/f7xiesnm/project-zipursky/data/03_flatfused' # 273LU_aldh1l1Cre_Pcdhg_homo_L23_cortex_autos1_flatfused_tile0.n5/'
outdir = '/u/home/f/f7xiesnm/project-zipursky/easifish/results/viz_273LU'
!mkdir -p $outdir

!ls $p

# raw


# rounds = ['r1', 'r2', 'r3', 'r4', 'r5']

rounds = np.arange(7) # ', 'r2', 'r3', 'r4', 'r5']
for r in rounds:
    path = p + f'/273LU_aldh1l1Cre_Pcdhg_homo_L23_cortex_autos1_flatfused_tile{r}.n5'
    zarr_data = zarr.open(store=zarr.N5Store(path), mode='r')
    for c in ['c0', 'c1', 'c2', 'c3', 'c4']:
        # img = zarr_data[f'/{c}/s4'] 
        # print(np.max(img), np.min(img))
        # output = outdir+f'/{r}_{c}_s4.tiff'
        # print(output)
        # tifffile.imwrite(output, img)
        
        img = zarr_data[f'/{c}/s2'] 
        # print(np.max(img), np.min(img))
        output = outdir+f'/{r}_{c}_s2.tiff'
        print(output)
        tifffile.imwrite(output, img)

rounds = [1] #  np.arange(7) # ', 'r2', 'r3', 'r4', 'r5']
for r in rounds:
    path = p + f'/273LU_aldh1l1Cre_Pcdhg_homo_L23_cortex_autos1_flatfused_tile{r}.n5'
    zarr_data = zarr.open(store=zarr.N5Store(path), mode='r')
    for c in ['c0', 'c1', 'c2', 'c3', 'c4']:
        # img = zarr_data[f'/{c}/s4'] 
        # print(np.max(img), np.min(img))
        # output = outdir+f'/{r}_{c}_s4.tiff'
        # print(output)
        # tifffile.imwrite(output, img)
        
        img = zarr_data[f'/{c}/s1'] 
        # print(np.max(img), np.min(img))
        output = outdir+f'/{r}_{c}_s1.tiff'
        print(output)
        tifffile.imwrite(output, img)
