==> analysis_easifish/bydatasets/273LU_L23/05.co-analyze-240611.ipynb <==
# analysis V4 - new basis code - Dec 6, 2023
- cleaned up the data QC and organizations a bit
- having two anndata and their spin offs - both FISH and proj data 
- flip an axis (y-axis) when plotting - not in data
- focus on L2/3 cells

TODO: 
- separate and organize plotting functions 
- organize the plots and generate more insights

import numpy as np
import pandas as pd
import os 
import matplotlib.pyplot as plt
import umap
import seaborn as sns
from sklearn.decomposition import PCA
from sklearn.cluster import KMeans
import matplotlib

import anndata 
import scanpy as sc
from scipy.stats import spearmanr

from scroutines import config_plots
from scroutines import basicu
from scroutines import miscu

import sys

# expr_nrm0 = np.loadtxt('/u/home/f/f7xiesnm/project-zipursky/easifish/296LURU/proc/batch_norm_counts.txt')
# expr_nrm1 = np.loadtxt('/u/home/f/f7xiesnm/project-zipursky/easifish/273LU/proc/batch_norm_counts.txt')
# expr_nrm2 = np.loadtxt('/u/home/f/f7xiesnm/project-zipursky/easifish/296LU/proc/batch_norm_counts.txt')
# expr_nrm3 = np.loadtxt('/u/home/f/f7xiesnm/project-zipursky/easifish/456RU/proc/batch_norm_counts.txt')

# exprs_nrm = [
#     expr_nrm0,
#     expr_nrm1,
#     expr_nrm2,
#     expr_nrm3,
# ]

expr0 = np.loadtxt('/u/home/f/f7xiesnm/project-zipursky/easifish/296LURU/proc/norm_counts.txt')
expr1 = np.loadtxt('/u/home/f/f7xiesnm/project-zipursky/easifish/273LU/proc/norm_counts.txt')
expr2 = np.loadtxt('/u/home/f/f7xiesnm/project-zipursky/easifish/296LU/proc/norm_counts.txt')
expr3 = np.loadtxt('/u/home/f/f7xiesnm/project-zipursky/easifish/456RU/proc/norm_counts.txt')

exprs = [
    expr0,
    expr1,
    expr2,
    expr3,
]

titles = ['296LURU\nwild type', '273LU\nKO', '296LU\nKO + Pcdhgc3 OE', '456RU\n KO + Pcdhga1 OE']

set_slc1a3 = 100

exprs_nrm = [
    expr/(np.median(expr[:,1])/set_slc1a3)
    for expr in exprs
]


for i, expr in enumerate(exprs):
    r, k, g = expr[:,0], expr[:,1], expr[:,2]
    sns.histplot(k, element='step', label=i, stat='density')
plt.legend()
plt.show()

for i, expr_nrm in enumerate(exprs_nrm):
    r, k, g = expr_nrm[:,0], expr_nrm[:,1], expr_nrm[:,2]
    sns.histplot(k, element='step', label=i, stat='density')
plt.legend()
plt.show()

print(np.sum(k > 200)/len(k))
print(np.sum(k > 210)/len(k))
print(np.sum(k > 220)/len(k))
print(np.sum(k > 230)/len(k))
cond_astro = k > 210

fig, axs = plt.subplots(1,4,figsize=(4*5,1*5), sharex=True, sharey=True)
for i, expr in enumerate(exprs_nrm):
    title = titles[i]
    ax = axs[i]
    ax.set_title(title)
    r, k, g = expr[:,0], expr[:,1], expr[:,2]
    ax.scatter(k, g, s=5)
    # ax.scatter(k[cond_astro], g[cond_astro], s=5, label='astro')
    ax.set_xlabel('Slc1a3')
    ax.set_ylabel('GFP')
# ax.set_xlim([0,400])
plt.show()

fig, axs = plt.subplots(1,4,figsize=(4*5,1*5), sharex=True, sharey=True)
for i, expr in enumerate(exprs_nrm):
    ax = axs[i]
    title = titles[i]
    ax.set_title(title)
    r, k, g = expr[:,0], expr[:,1], expr[:,2]
    ax.scatter(k, r, s=5)
    # ax.scatter(k[cond_astro], r[cond_astro], s=5, label='astro')
    ax.set_ylabel('mCherry')
    ax.set_xlabel('Slc1a3')
    
# ax.set_xlim([0,400])
plt.show()

fig, axs = plt.subplots(1,4,figsize=(4*5,1*5), sharex=True, sharey=True)
for i, expr in enumerate(exprs_nrm):
    title = titles[i]
    ax = axs[i]
    ax.set_title(title)
    r, k, g = expr[:,0], expr[:,1], expr[:,2]
    ax.scatter(r, g, s=5)
    # ax.scatter(k[cond_astro], g[cond_astro], s=5, label='astro')
    ax.set_xlabel('mCherry')
    ax.set_ylabel('GFP')
plt.show()

fig, ax = plt.subplots(1,1,figsize=(1*5,1*5)) # , sharex=True, sharey=True)
i = 1
expr = exprs_nrm[i]
title = titles[i]
# ax = axs[i]
ax.set_title(title)
r, k, g = expr[:,0], expr[:,1], expr[:,2]
ax.scatter(k, g, s=5)
# ax.scatter(k[cond_astro], g[cond_astro], s=5, label='astro')
ax.set_xlabel('Slc1a3')
ax.set_ylabel('GFP')
# ax.set_xlim([0,400])

fig, ax = plt.subplots(1,1,figsize=(1*5,1*5)) # , sharex=True, sharey=True)
i = 3
expr = exprs_nrm[i]
title = titles[i]
# ax = axs[i]
ax.set_title(title)
r, k, g = expr[:,0], expr[:,1], expr[:,2]
ax.scatter(k, g, s=5)
# ax.scatter(k[cond_astro], g[cond_astro], s=5, label='astro')
ax.set_xlabel('Slc1a3')
ax.set_ylabel('GFP')
# ax.set_xlim([0,400])
